requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-unfold==0.12.0
